requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
Faker==15.1.1
mysqlclient==2.1.1
numpy==1.23.4
pandas==1.5.1
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
